moved package pyslate2 to pyslate
