+++ pyslate/cache.py
