+++ pyslate/locales.py
